requirements.txt
albumentations==1.3.1
matplotlib==3.6.0
numpy==1.23.3
opencv_python==4.6.0.66
Pillow==10.0.0
pytorch_lightning==2.0.5
scikit_learn==1.3.0
torch==1.12.1+cu116
